pyproject.toml
[tool.poetry]
name = "server_practice"
version = "0.0.1"
description = ""
authors = ["smapl <a.mustafin00@mail.ru>"]


[tool.poetry.dependencies]
python = "^3.6.6"
flask-restplus = "^0.13.0"
requests = "^2.23.0"
psycopg2-binary = "^2.8.4"



[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
